Data structure change: removing use of dictionary
A dictionary structure for mantaining the books to sort won't always work,
because a book list might have at least two different books with
identical titles (or identical authors/edition year). That case only one
would be stored in the dictionary.

Using the optional parameter 'key' of 'sorted()' built-in function, it
is possible to use a list of type Book and select the specific parameter
which is the criterion to sort.