Using abstract class BookSorter; sublists calculation
Reducing amout of code by making use of BookSorter as a really Abstract
Class, not just as a Interface.

The key on this is the use of the optional parameter 'key' of built-in
function 'sorted()'.

Now the sublists that have to be sorted accordingly the next Strategy
(sorted by other book parameter) are calculated by equal_elements()
abstract method.